+++ docs/book/src/cronjob-tutorial/testdata/project/config/default/metrics_service.yaml
