+++ docs/book/src/getting-started/testdata/project/config/default/metrics_service.yaml
