+++ testdata/project-v4-multigroup-with-deploy-image/config/default/metrics_service.yaml
