+++ testdata/project-v4-multigroup/config/default/metrics_service.yaml
