+++ testdata/project-v4-with-deploy-image/config/default/metrics_service.yaml
