+++ testdata/project-v4-with-grafana/config/default/metrics_service.yaml
