+++ testdata/project-v4-with-grafana/config/rbac/kustomization.yaml
@@ -9,4 +9,3 @@ resources:
 - role_binding.yaml
 - leader_election_role.yaml
 - leader_election_role_binding.yaml
-- metrics_service.yaml
