+++ testdata/project-v4/config/default/metrics_service.yaml
